requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
appdirs==1.4.3            # via fs
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
billiard==3.3.0.23        # via celery
celery==3.1.18
coverage==4.5.1           # via pytest-cov
ddt==1.1.3
django-braces==1.13.0     # via django-oauth-toolkit
django-model-utils==3.1.2
django-oauth-toolkit==0.12.0
djangorestframework==3.6.4
edx-celeryutils==0.1.5
edx-completion==0.1.7
edx-opaque-keys[django]==0.4.4
fs==2.0.24                # via xblock
jsonfield==2.0.2          # via edx-celeryutils
kombu==3.0.37             # via celery
lxml==4.2.3               # via xblock
markupsafe==1.0           # via xblock
mock==2.0.0
more-itertools==4.2.0     # via pytest
mysqlclient==1.3.13
oauthlib==2.0.1           # via django-oauth-toolkit
pbr==4.1.0                # via mock, stevedore
pluggy==0.6.0             # via pytest
py==1.5.4                 # via pytest
pymongo==3.7.0            # via edx-opaque-keys
pytest-cov==2.5.1
pytest==3.6.3             # via pytest-cov, pytest-django
python-dateutil==2.7.3    # via xblock
pytz==2018.5              # via celery, edx-completion, fs, xblock
pyyaml==3.13              # via xblock
redis==2.10.6
six==1.11.0
stevedore==1.28.0         # via edx-opaque-keys
web-fragments==0.2.2      # via xblock
webob==1.8.2              # via xblock
xblock==1.2.1
